backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Vector database
chromadb==0.4.18

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# PDF processing
PyMuPDF==1.23.8

# Authentication
PyJWT==2.8.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Email validation
email-validator==2.1.0

# Monitoring and logging
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7
structlog==23.2.0

# AI services (optional - add API keys to .env)
openai==1.3.7
google-generativeai==0.3.1
